+++ src/View/Announcement/category.html.twig
@@ -1,0 +1,11 @@
+{% extends 'layout.html.twig' %}
+
+{% block content %}
+    <div class="container">
+    {% include 'Components/_cart.html.twig' %}
+    {% include 'Components/_cart.html.twig' %}
+    {% include 'Components/_cart.html.twig' %}
+    {% include 'Components/_cart.html.twig' %}
+    {% include 'Components/_cart.html.twig' %}
+    </div>    
+{% endblock %}
