+++ src/View/Components/_header.html.twig
@@ -9,7 +9,7 @@
 			</a>
 			<ul class="ul_header">
 				<li>
-					<a href="/concept" class="a_header">Le concept</a>
+					<a href="/about" class="a_header">Le concept</a>
 				</li>
 				<li>
 					<a href="/connexion" class="a_header">Connexion</a>
@@ -25,9 +25,9 @@
 		<img src="/assets/images/icons/search_icone.png" alt="search_icone" class="search_icone"/>
 		<input class="searchbar" onkeyup="search_GooDeal()" type="text" name="search" placeholder="Rechercher un GooDeal...">
 	</div>
-	<a href="/about" class="a_header">Le concept</a>
+	<a href="/about" class="a_header about">Le concept</a>
 </header>
 
 {% block javascript %}
-	<script src="./assets/javascript/header.js"></script>
+	<script src="/assets/javascript/header.js"></script>
 {% endblock %}
