+++ src/View/Home/index.html.twig
@@ -27,19 +27,17 @@
                  <div class="lastNews">
                     <h2>Vos derniers avis</h2>
                     <button class="btnGooDeals"
-                      type="button" onclick="location.href='../items/category'" >
+                      type="button" onclick="location.href='../announcement/category'" >
                       Ajoute ton GooDeals</a>
                     </button>
                     <div class="allArrows">
-                    <img src="assets/images/leftArrow.svg" class="leftArrow">
+                    <img src="assets/images/icons/leftArrow.svg" class="leftArrow">
                       <div class="slide">
                       <div class="card card1"><div>
                       <div class="card card2"></div>
                       <div class="card card3"></div>
-                    <img src="assets/images/rightArrow.svg" class="rightArrow">
-
-                    </div>
-                 
+                    <img src="assets/images/icons/rightArrow.svg" class="rightArrow">
+                    </div>                 
                 </div>
             </section>
         </main>
