--- src/View/Item/category.html.twig
@@ -1,11 +1,0 @@
-{% extends 'layout.html.twig' %}
-
-{% block content %}
-    <div class="container">
-    {% include 'Item/_cart.html.twig' %}
-    {% include 'Item/_cart.html.twig' %}
-    {% include 'Item/_cart.html.twig' %}
-    {% include 'Item/_cart.html.twig' %}
-    {% include 'Item/_cart.html.twig' %}
-    </div>    
-{% endblock %}
